+++ src/commands/migrate_packages.rs
@@ -81,7 +81,6 @@ pub(crate) fn run(opt: OptCompeteMigratePackages, ctx: crate::Context<'_>) -> an
         .collect::<Result<Vec<_>, ignore::Error>>()?;
 
     let mut include = vec![];
-    let mut submit_via_binary = false;
 
     for manifest_path in manifest_paths.into_iter().sorted() {
         let metadata = crate::project::cargo_metadata_no_deps_frozen(&manifest_path)?;
@@ -104,8 +103,6 @@ pub(crate) fn run(opt: OptCompeteMigratePackages, ctx: crate::Context<'_>) -> an
     for package in &include {
         let mut manifest =
             crate::fs::read_to_string(&package.manifest_path)?.parse::<toml_edit::Document>()?;
-
-        submit_via_binary |= !manifest["profile"].is_none();
 
         manifest["profile"] = toml_edit::Item::None;
 
@@ -159,8 +156,42 @@ pub(crate) fn run(opt: OptCompeteMigratePackages, ctx: crate::Context<'_>) -> an
         shell,
     )?;
 
+    let cargo_atcoder_config = (|| -> _ {
+        let path = dirs::config_dir()?.join("cargo-atcoder.toml");
+        crate::fs::read_to_string(path)
+            .ok()?
+            .parse::<toml_edit::Document>()
+            .ok()
+    })();
+
+    let submit_via_binary = matches!(
+        &cargo_atcoder_config,
+        Some(c) if c["atcoder"]["submit_via_binary"].as_bool() == Some(true)
+    );
+
+    let mut root_manifest = r#"[workspace]
+members = []
+exclude = []
+"#
+    .parse::<toml_edit::Document>()
+    .unwrap();
+
+    if submit_via_binary {
+        if let Some(profile_release) = (|| -> _ {
+            let path = dirs::config_dir()?.join("cargo-atcoder.toml");
+            let config = crate::fs::read_to_string(path)
+                .ok()?
+                .parse::<toml_edit::Document>()
+                .ok()?;
+            Some(config["profile"]["release"].clone())
+        })() {
+            root_manifest["profile"] = implicit_table();
+            root_manifest["profile"]["release"] = profile_release;
+        }
+    }
+
     let root_manifest_path = path.join("Cargo.toml");
-    crate::fs::write(&root_manifest_path, "[workspace]\n")?;
+    crate::fs::write(&root_manifest_path, root_manifest.to_string())?;
     shell.status("Wrote", root_manifest_path.display())?;
 
     shell.status("Adding", format!("{} + 1 packages", include.len()))?;