+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [`compete.toml`] Changed the behavior of `exclude`.
+
+### Added
+
+- [`compete.toml`] `skip` variant to `new-workspace-members`.
+
+### Fixed
+
+- [`migrate` command] Now `cargo compete migrate` also generates a template package as `cargo compete init` does.
+
 ## [0.3.2] - 2020-08-08Z
 
 ### Added